mnist_classifier/requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
Pillow
psycopg2-binary